QPUConnection now takes device, not device_name, as argument. Modified tests required moving some fixtures from test_device.py to a new conftest.py file.
